Move _extract_direction to DL proper
