+++ src/styles.jl
@@ -385,3 +385,17 @@ to_polygons(ent::GeometryEntity, ::WithDirection; kwargs...) = to_polygons(ent; 
 
 transform(sty::WithDirection, f::Transformation) =
     WithDirection(rotated_direction(sty.direction, f))
+
+# Walk through any nesting of StyledEntity wrappers and return the `direction`
+# angle of the first `WithDirection` style encountered (from outside in), or `nothing` if no
+# `WithDirection` is present. Handles nesting like
+# WithDirection(MeshSized(only_simulated(rect))) and the reverse.
+_extract_direction(::DeviceLayout.GeometryEntity) = nothing
+function _extract_direction(ent::DeviceLayout.StyledEntity)
+    return _extract_direction(ent.ent)
+end
+function _extract_direction(
+    ent::DeviceLayout.StyledEntity{T, U, WithDirection}
+) where {T, U <: GeometryEntity{T}}
+    return ent.sty.direction
+end
